클라이언트용 메시지 코드와 클러스터 노드용 메시지 코드 통합
