+++ src/main/java/databute/databuter/client/register/RegisterMessage.java
@@ -1,7 +1,6 @@
 package databute.databuter.client.register;
 
 import com.google.common.base.MoreObjects;
-import databute.databuter.client.network.ClientMessageCode;
 import databute.databuter.network.message.Message;
 import databute.databuter.network.message.MessageCode;
 
@@ -9,7 +8,7 @@ public class RegisterMessage implements Message {
 
     @Override
     public MessageCode messageCode() {
-        return ClientMessageCode.REGISTER;
+        return MessageCode.REGISTER;
     }
 
     @Override
