+++ src/main/java/databute/databuter/cluster/handshake/response/HandshakeResponseMessage.java
@@ -1,7 +1,6 @@
 package databute.databuter.cluster.handshake.response;
 
 import com.google.common.base.MoreObjects;
-import databute.databuter.cluster.network.ClusterMessageCode;
 import databute.databuter.network.message.Message;
 import databute.databuter.network.message.MessageCode;
 
@@ -9,7 +8,7 @@ public class HandshakeResponseMessage implements Message {
 
     @Override
     public MessageCode messageCode() {
-        return ClusterMessageCode.HANDSHAKE_RESPONSE;
+        return MessageCode.HANDSHAKE_RESPONSE;
     }
 
     @Override
