+++ src/main/java/databute/databuter/cluster/notification/ClusterNodeNotificationMessage.java
@@ -1,7 +1,6 @@
 package databute.databuter.cluster.notification;
 
 import com.google.common.base.MoreObjects;
-import databute.databuter.client.network.ClientMessageCode;
 import databute.databuter.network.message.Message;
 import databute.databuter.network.message.MessageCode;
 
@@ -31,7 +30,7 @@ public class ClusterNodeNotificationMessage implements Message {
 
     @Override
     public MessageCode messageCode() {
-        return ClientMessageCode.CLUSTER_NODE_NOTIFICATION;
+        return MessageCode.CLUSTER_NODE_NOTIFICATION;
     }
 
     public ClusterNodeNotificationType type() {
